% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_ref.R
\name{add_fig}
\alias{add_fig}
\alias{add_sfig}
\title{Add a figure}
\usage{
add_fig(fig_path, ref_id, caption_main, caption_text = "", width = 5,
  height = 7, format = get_default_format())

add_sfig(fig_path, ref_id, caption_main, caption_text = "", width = 5,
  height = 7, format = get_default_format(), integrate = TRUE)
}
\arguments{
\item{fig_path}{Path to the figure}

\item{ref_id}{The identifier}

\item{caption_main}{Caption title}

\item{caption_text}{Caption text}

\item{width}{Width in inches}

\item{height}{Height in inches}

\item{format}{The format, in html or latex}
}
\description{
Add a figure
}
