% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_benchmark_aggregate.R
\name{benchmark_aggregate}
\alias{benchmark_aggregate}
\title{Normalisation and aggregation function}
\usage{
benchmark_aggregate(data, metrics = c("correlation", "him",
  "featureimp_cor", "F1_branches"), norm_fun = "normal",
  mean_fun = "geometric", mean_weights = set_names(rep(1,
  length(metrics)), metrics),
  dataset_source_weights = get_default_dataset_weights())
}
\arguments{
\item{data}{Data as generated by the `06-benchmark/2-retrieve_results.R`` script}

\item{metrics}{Metrics to normalise}

\item{norm_fun}{Which normalisation to use}

\item{mean_fun}{Which mean function to use}

\item{mean_weights}{Which metrics to use for the calculation of the mean, and which weights}

\item{dataset_source_weights}{The weights of the dataset sources}
}
\description{
Normalisation and aggregation function
}
