% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_benchmark_fetch_results.R
\name{benchmark_bind_results}
\alias{benchmark_bind_results}
\title{Gather and bind the results of the benchmark jobs}
\usage{
benchmark_bind_results(load_models = FALSE)
}
\arguments{
\item{load_models}{Whether or not to load the models as well.}
}
\description{
Gather and bind the results of the benchmark jobs
}
