% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_benchmark_fetch_results.R
\name{benchmark_fetch_results}
\alias{benchmark_fetch_results}
\title{Fetch the results of the benchmark jobs from the cluster.}
\usage{
benchmark_fetch_results()
}
\description{
Fetch the results of the benchmark jobs from the cluster.
}
