% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_benchmark_submit.R
\name{benchmark_submit}
\alias{benchmark_submit}
\title{A benchmark suite with which to run all the methods on the different datasets}
\usage{
benchmark_submit(design, timeout_per_execution = 3600,
  max_memory_per_execution = "10G", metrics = "correlation",
  local_output_folder, remote_output_folder, execute_before = NULL,
  verbose = TRUE)
}
\arguments{
\item{design}{Design tibble of the experiment, created by \code{\link[=benchmark_generate_design]{benchmark_generate_design()}}.}

\item{timeout_per_execution}{The maximum number of seconds each execution is allowed to run, otherwise it will fail.}

\item{max_memory_per_execution}{The maximum amount of memory each execution is allowed to use, otherwise it will fail.}

\item{metrics}{Which metrics to evaluate; see \code{\link[=calculate_metrics]{calculate_metrics()}} for a list of which metrics are available.}

\item{local_output_folder}{A folder in which to output intermediate and final results.}

\item{remote_output_folder}{A folder in which to store intermediate results in a remote directory when using the qsub package.}

\item{execute_before}{Shell commands to execute before running R.}

\item{verbose}{Whether or not to print extra information.}
}
\description{
A benchmark suite with which to run all the methods on the different datasets
}
