% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{data}
\name{characterised_metrics}
\alias{characterised_metrics}
\title{Metrics used in the characterisation}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 7 columns.}
\usage{
characterised_metrics
}
\description{
Metrics used in the characterisation
}
\keyword{datasets}
