% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg_replacement.R
\name{create_replacers}
\alias{create_replacers}
\title{Create replacers
Adds a replace_id column}
\usage{
create_replacers(to_replace)
}
\arguments{
\item{to_replace}{Dataframe containing the data}
}
\description{
Create replacers
Adds a replace_id column
}
