% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_loading.R
\name{dataset_location}
\alias{dataset_location}
\title{Determining the location of a dataset after it has been preprocessed}
\usage{
dataset_location(prefix, dataset_id)
}
\arguments{
\item{prefix}{Dataset prefix}

\item{dataset_id}{Dataset id}
}
\description{
Determining the location of a dataset after it has been preprocessed
}
