% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_loading.R
\name{dataset_preproccessing}
\alias{dataset_preproccessing}
\alias{dataset_preproc_file}
\alias{dataset_file}
\title{#' @export
load_dataset <- function(dataset_id, prefix="") {
  if(file.exists("/group/irc/shared")) {
    read_rds(paste0("/group/irc/shared/dynalysis/analysis/data/datasets/", prefix, "/", dataset_id, ".rds"))
  } else {
    read_rds(paste0("analysis/data/datasets/", prefix, "/", dataset_id, ".rds"))
  }
}}
\usage{
dataset_preproccessing(prefix, dataset_id)

dataset_preproc_file(...)

dataset_file(...)
}
\arguments{
\item{prefix}{Dataset prefix}

\item{dataset_id}{Dataset id}
}
\description{
#' @export
save_dataset <- function(dataset) {
  write_rds(dataset, paste0("analysis/data/datasets/real/", dataset$info$id, ".rds"))
}
Helper function for creating new datasets
}
