% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dataset_preprocessing.R
\name{datasetpreproc_normalise_filter_wrap_and_save}
\alias{datasetpreproc_normalise_filter_wrap_and_save}
\title{Preprocessing functionality for real datasets}
\usage{
datasetpreproc_normalise_filter_wrap_and_save(counts, cell_ids, milestone_ids,
  milestone_network, milestone_percentages, cell_grouping, cell_info,
  feature_info, dataset_id = NULL)
}
\arguments{
\item{counts}{The counts.}

\item{cell_ids}{The ids of the cells.}

\item{milestone_ids}{The ids of the milestones in the trajectory. Type: Character vector.}

\item{milestone_network}{The network of the milestones.
Type: Data frame(from = character, to = character, length = numeric, directed = logical).}

\item{milestone_percentages}{A data frame specifying what percentage milestone each cell
consists of.
Type: Data frame(cell_id = character, milestone_id = character, percentage = numeric).}

\item{cell_grouping}{Milestone groups of the cells.}

\item{cell_info}{Optional meta-information pertaining the cells.}

\item{feature_info}{Optional meta-information pertaining the features.}

\item{dataset_id}{The name of the dataset.}
}
\description{
Preprocessing functionality for real datasets
}
