% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dataset_preprocessing.R
\name{download_dataset_file}
\alias{download_dataset_file}
\title{Download a file and return its location path}
\usage{
download_dataset_file(url, filename)
}
\arguments{
\item{url}{The url of the file to download}

\item{filename}{What name to give to the file}
}
\description{
Download a file and return its location path
}
