% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{dynalysis}
\alias{dynalysis}
\alias{dynalysis-package}
\title{Easy installing and loading of the dynverse}
\description{
The 'dynalysis' is a set of packages with which to
evaluate trajectory inference methods for dynamic processes.
}
