% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{error_reasons}
\alias{error_reasons}
\title{Metadata on error reasons}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 3 columns.}
\usage{
error_reasons
}
\description{
Metadata on error reasons
}
\keyword{datasets}
