% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_method_characterisation.R
\docType{data}
\name{errors}
\alias{errors}
\title{Errors}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 3 columns.}
\usage{
errors
}
\description{
Errors
}
\keyword{datasets}
