% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_platforms.R
\name{estimate_platform}
\alias{estimate_platform}
\title{Estimate a platform}
\usage{
estimate_platform(dataset_id, subsample = 500, override_fun = FALSE)
}
\arguments{
\item{dataset_id}{The dataset_id from which the platform will be estimated, using the files in datasets_preproc/raw}

\item{subsample}{The number of cells to subsample}

\item{override_fun}{Whether or not to override one splatEstDropout params}
}
\description{
Estimate a platform
}
