% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{data}
\name{evaluated_metrics}
\alias{evaluated_metrics}
\title{Metrics used in the evaluation}
\format{An object of class \code{character} of length 8.}
\usage{
evaluated_metrics
}
\description{
Metrics used in the evaluation
}
\keyword{datasets}
