% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{data}
\name{evaluation_metrics}
\alias{evaluation_metrics}
\title{Metrics used in the evaluation}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 7 columns.}
\usage{
evaluation_metrics
}
\description{
Metrics used in the evaluation
}
\keyword{datasets}
