% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_benchmark_fetch_results.R
\name{extract_error_status}
\alias{extract_error_status}
\title{Extract the error status from an evaluation}
\usage{
extract_error_status(model, stdout, stderr, error_message, ...)
}
\arguments{
\item{model}{The model}

\item{stdout}{The standard output}

\item{stderr}{The standard error}

\item{error_message}{The error message}

\item{...}{Other columns in the output, ignored}
}
\description{
Extract the error status from an evaluation
}
