% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_document_scripts.R
\name{extract_scripts_documentation}
\alias{extract_scripts_documentation}
\alias{render_scripts_documentation}
\title{Will recursively extract documentations from the scripts and folders within a given folder}
\usage{
extract_scripts_documentation(folder = getwd(), recursive = TRUE)

render_scripts_documentation(folder = ".", recursive = FALSE)
}
\arguments{
\item{folder}{The folder from which to start, defaults to current working directory}

\item{recursive}{Whether to extract scripts recursively}
}
\description{
Will recursively extract documentations from the scripts and folders within a given folder
}
