% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formats.R
\name{fix_relative_paths}
\alias{fix_relative_paths}
\title{process relative paths to links & figures
match every link and check whether it is an absolute path (= starts with / or http) or not
every relative path gets updated}
\usage{
fix_relative_paths(knit, folder)
}
\description{
process relative paths to links & figures
match every link and check whether it is an absolute path (= starts with / or http) or not
every relative path gets updated
}
\examples{
knit <- c(
"hshlkjdsljkfdhg [xxxxx](/pompompom/dhkjhlkj/) kjfhlqkjsdhlkfjqsdf",
"hshlkjdsljkfdhg [xxxxx](pompompom/dhkjhlkj/) kjfhlqkjsdhlkfjqsdf",
"<img src = \\"heyho/heyho\\">",
"<img src = \\"/heyho/heyho\\">"
)
fix_relative_links(knit, "IT WORKED :)")
}
