% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funky_heatmap.R
\name{funky_heatmap}
\alias{funky_heatmap}
\title{Overview heatmap plotting}
\usage{
funky_heatmap(data, metric_ord, method_ord, add_timestamp = TRUE,
  row_space = 0.1, row_bigspace = 0.5, col_space = 0.1,
  col_bigspace = 0.5)
}
\arguments{
\item{data}{data}

\item{metric_ord}{metric_ord}

\item{method_ord}{method_ord}

\item{add_timestamp}{Whether or not to add a timestamp at the bottom}

\item{row_space}{The space between rows}

\item{row_bigspace}{The space between groups of rows}

\item{col_space}{The space between columns}

\item{col_bigspace}{The space between groups of columns}
}
\description{
TODO: params need to be described
}
