% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_benchmark.R
\name{generate_benchmark_design}
\alias{generate_benchmark_design}
\title{Generate a benchmarking design}
\usage{
generate_benchmark_design(dataset_ids, method_ids,
  parameters = map(method_ids, ~tibble(paramset_ix = 1)),
  give_priors = NULL, num_repeats = 1)
}
\arguments{
\item{dataset_ids}{The ids of the datasets to be used in the evaluation.}

\item{method_ids}{The ids of the methods to be evaluated.}

\item{parameters}{A named list containing data frames of the parameters to evaluate.
The names of the list must be present in method_ids.
The data frames must correspond to the parameter set \code{par_set} of each method,
e.g. as generated by \code{\link[ParamHelpers]{generateDesignOfDefaults}},
but must also contain the column \code{paramset_ix}.}

\item{give_priors}{All the priors a method is allowed to receive. Must be a subset of: \code{"start_milestones"},
\code{"start_id"}, \code{"end_milestones"}, \code{"end_id"}, \code{"groups_id"} and \code{"groups_network"}}

\item{num_repeats}{The number of times to repeat the evaluation.}
}
\description{
Generate a benchmarking design
}
\examples{
generate_benchmark_design(
  dataset_ids = c("toy/bifurcating_1", "toy/bifurcating_2"),
  method_ids = c("angle", "scorpius", "tscan"),
  parameters = list(scorpius = tibble(paramset_ix = 1), tscan = tibble(paramset_ix = 1:3, clusternum_lower = 4:6, clusternum_upper = 18:20)),
  give_priors = NULL,
  num_repeats = 2
)
}
