% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_priors_mini.R
\name{generate_prior_mini}
\alias{generate_prior_mini}
\title{Creates a prior mini svg}
\usage{
generate_prior_mini(method_priors,
  svg = xml2::read_xml(figure_file("priors.svg", experiment_id = "priors")))
}
\arguments{
\item{method_priors}{Dataframe containing a prior_id and a prior_usage}

\item{svg}{The base svg of the priors}
}
\description{
Creates a prior mini svg
}
