% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dynalysis_folder.R
\name{get_dynalysis_folder}
\alias{get_dynalysis_folder}
\title{Find the dynalysis path}
\usage{
get_dynalysis_folder()
}
\description{
When executing some of the dynalysis functions, the git repository needs to be found.
Either set the working directory to the repository, or define the DYNALYSIS_PATH variable.
}
