% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_get_dynbenchmark_folder.R
\name{get_dynbenchmark_folder}
\alias{get_dynbenchmark_folder}
\title{Find the dynbenchmark path}
\usage{
get_dynbenchmark_folder()
}
\description{
When executing some of the dynbenchmark functions, the git repository needs to be found.
Either set the working directory to the repository, or define the DYNBENCHMARK_PATH variable.
}
