% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formats.R
\name{github_markdown_nested}
\alias{github_markdown_nested}
\title{A nested github markdown document}
\usage{
github_markdown_nested(...)
}
\arguments{
\item{...}{Parameters for rmarkdown::github_document}
}
\description{
A nested github markdown document
}
