% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_scholar.R
\name{google_scholar}
\alias{google_scholar}
\title{Find the number of citations for a given google scholar id}
\usage{
google_scholar(cluster_id)
}
\arguments{
\item{cluster_id}{The Google Scholar cluster id to query}
}
\description{
Find the number of citations for a given google scholar id
}
