% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headtail.R
\name{headtail}
\alias{headtail}
\title{Retain the heads and tails of a character vector split by newlines}
\usage{
headtail(X, num)
}
\arguments{
\item{X}{A character vector to process}

\item{num}{The number of lines to retain at the head and at the tail}
}
\description{
Retain the heads and tails of a character vector split by newlines
}
