% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{id_mapper}
\alias{id_mapper}
\title{Mapping ensembl or entrez ids to symbols}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 129094 rows and 3 columns.}
\usage{
id_mapper
}
\description{
Mapping ensembl or entrez ids to symbols
}
\keyword{datasets}
