% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_document_scripts.R
\name{knit_child_readme}
\alias{knit_child_readme}
\title{Knit a child README, and add an extra level of headings + fix relative paths}
\usage{
knit_child_readme(folder)
}
\arguments{
\item{folder}{Subfolder where the README.Rmd is located}
}
\description{
Knit a child README, and add an extra level of headings + fix relative paths
}
