% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_labels.R
\name{label_facet}
\alias{label_facet}
\title{Labeller for facets}
\usage{
label_facet(label_func = label_long, ...)
}
\arguments{
\item{label_func}{Which function to use for facet labelling}

\item{...}{Passed to mutate_all}
}
\description{
Labeller for facets
}
