% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_labels.R
\name{label_memory}
\alias{label_memory}
\title{Label memory}
\usage{
label_memory(x, include_mb = FALSE)
}
\arguments{
\item{x}{Memory in bytes}

\item{include_mb}{Also include MB values.}
}
\description{
Label memory
}
