% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_labels.R
\name{label_metric}
\alias{label_metric}
\title{Label the metrics}
\usage{
label_metric(metric_id, parse = FALSE)
}
\arguments{
\item{metric_id}{metric id}

\item{parse}{Whether to parse the label into an expression}
}
\description{
Label the metrics
}
