% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_labels.R
\name{label_pvalue}
\alias{label_pvalue}
\title{Label p_value}
\usage{
label_pvalue(p_values, cutoffs = c(0.1, 0.01, 1e-05))
}
\arguments{
\item{p_values}{P values}

\item{cutoffs}{Cutoffs for number of asterisks}
}
\description{
Label p_value
}
