% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataraw_labels.R
\name{label_short}
\alias{label_short}
\title{Short labelling function}
\usage{
label_short(x, width = 10)
}
\arguments{
\item{x}{The text}

\item{width}{The width of the label}
}
\description{
Short labelling function
}
