% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_labels.R
\name{label_vector}
\alias{label_vector}
\title{Collapse a vector with commas and "ands"}
\usage{
label_vector(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Collapse a vector with commas and "ands"
}
