% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataraw_labels.R
\name{label_wrap}
\alias{label_wrap}
\title{Text wrapping}
\usage{
label_wrap(x, width = 10, collapse = "\\n")
}
\arguments{
\item{x}{a character vector.}

\item{width}{a positive integer giving the target column for wrapping lines in the output.}

\item{collapse}{an optional character string to separate the different lines.}
}
\description{
Text wrapping
}
