% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_themes.R
\name{legend_at}
\alias{legend_at}
\title{Only retain the legend in the middle plot}
\usage{
legend_at(plots, at = floor(length(plots)/2), theme_legend = theme())
}
\arguments{
\item{plots}{A list of plots}

\item{at}{Integers at which a legend will be added}

\item{theme_legend}{The theme to apply to the plot with a legend}
}
\description{
Only retain the legend in the middle plot
}
