% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_document_scripts.R
\name{link_to_results}
\alias{link_to_results}
\title{Generates a link to the results}
\usage{
link_to_results(scripts_folder = getwd())
}
\arguments{
\item{scripts_folder}{Folder from which to calculate the experiment id}
}
\description{
Generates a link to the results
}
