% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dataset_preprocessing.R
\name{load_dataset}
\alias{load_dataset}
\title{Load a dataset after it has been preprocessed}
\usage{
load_dataset(dataset_id = NULL)
}
\arguments{
\item{dataset_id}{The ID of the dataset to be used}
}
\description{
Load a dataset after it has been preprocessed
}
