% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dyneval_model.R
\name{load_dyneval_model}
\alias{load_dyneval_model}
\title{Load model from benchmark suite results}
\usage{
load_dyneval_model(method_id, df, auto_gc = TRUE, experiment_id = NULL)
}
\arguments{
\item{method_id}{The short_name of a method}

\item{df}{A data frame containing the crossing of which models to load}

\item{auto_gc}{Whether or not to automatically garbage collect after loading models}

\item{experiment_id}{The experiment id (default NULL}
}
\description{
Load model from benchmark suite results
}
