% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_methods.R
\name{load_methods}
\alias{load_methods}
\title{Load the information on the methods}
\usage{
load_methods(filter_fun = function(tib) tib \%>\% filter(source == "tool"
  | id \%in\% c("angle", "comp1 ", "mst")))
}
\arguments{
\item{filter_fun}{A function with which to filter the methods tibble.}
}
\description{
Load the information on the methods
}
