% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyneval_load_model.R
\name{load_model}
\alias{load_model}
\title{Load model from benchmark suite results}
\usage{
load_model(method_short_name, model_id, auto_gc = TRUE,
  experiment_id = NULL)
}
\arguments{
\item{method_short_name}{The short_name of a method}

\item{model_id}{The id of the model to be loaded}

\item{auto_gc}{Whether or not to automatically garbage collect after loading models}

\item{experiment_id}{The experiment id (default NULL}
}
\description{
Load model from benchmark suite results
}
