% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_pipeline.R
\name{load_or_generate}
\alias{load_or_generate}
\title{Load or generate an R object}
\usage{
load_or_generate(file_location, expression)
}
\arguments{
\item{file_location}{Location of the RDS file}

\item{expression}{Expression to run if the RDS file is not found}
}
\description{
Load or generate an R object
}
