% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_platforms.R
\name{load_platforms}
\alias{load_platforms}
\alias{load_platform}
\alias{get_platform_features}
\alias{select_platforms}
\title{List and load the platforms}
\usage{
load_platforms()

load_platform(platform_id)

get_platform_features(platforms)

select_platforms(n_platforms)
}
\arguments{
\item{platform_id}{Platform identifier}

\item{platforms}{The platforms, loaded by load_platforms}

\item{n_platforms}{Number of platforms to select}
}
\description{
List and load the platforms

Load a platform
}
