% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_paramoptim.R
\name{make_obj_fun}
\alias{make_obj_fun}
\title{Used for wrapping an evaluation function around a TI method}
\usage{
make_obj_fun(method, metrics, extra_metrics, noisy = FALSE, verbose = FALSE)
}
\arguments{
\item{noisy}{Whether or not the metric is noisy or not}

\item{verbose}{Whether or not to print extra information output}
}
\description{
Used for wrapping an evaluation function around a TI method
}
