% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{method_status_colours}
\alias{method_status_colours}
\title{Metadata on method statuses}
\format{An object of class \code{character} of length 7.}
\usage{
method_status_colours
}
\description{
Metadata on method statuses
}
\keyword{datasets}
