% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{method_statuses}
\alias{method_statuses}
\title{Metadata on method statuses}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 3 columns.}
\usage{
method_statuses
}
\description{
Metadata on method statuses
}
\keyword{datasets}
