% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{data}
\name{metrics_characterised}
\alias{metrics_characterised}
\title{Metrics used in the characterisation}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13 rows and 9 columns.}
\usage{
metrics_characterised
}
\description{
Metrics used in the characterisation
}
\keyword{datasets}
