% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{data}
\name{metrics_evaluated}
\alias{metrics_evaluated}
\title{Metrics used in the evaluation}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 10 columns.}
\usage{
metrics_evaluated
}
\description{
Metrics used in the evaluation
}
\keyword{datasets}
