% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{non_inclusion_reasons}
\alias{non_inclusion_reasons}
\title{Metadata on non-inclusion reasons}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 3 columns.}
\usage{
non_inclusion_reasons
}
\description{
Metadata on non-inclusion reasons
}
\keyword{datasets}
