% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_paramoptim.R
\name{paramoptim_bind_results}
\alias{paramoptim_bind_results}
\title{Gather and bind the results of the paramoptim jobs}
\usage{
paramoptim_bind_results(local_output_folder)
}
\arguments{
\item{local_output_folder}{The folder in which to output intermediate and final results.}
}
\description{
Gather and bind the results of the paramoptim jobs
}
