% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_paramoptim.R
\name{paramoptim_fetch_results}
\alias{paramoptim_fetch_results}
\title{Fetch the results of the paramoptim jobs from the cluster.}
\usage{
paramoptim_fetch_results(local_output_folder)
}
\arguments{
\item{local_output_folder}{The folder in which to output intermediate and final results.}
}
\description{
Fetch the results of the paramoptim jobs from the cluster.
}
