% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_paramoptim_generate_design.R
\name{paramoptim_generate_design}
\alias{paramoptim_generate_design}
\title{Generate a benchmarking design}
\usage{
paramoptim_generate_design(datasets, methods, priors = NULL,
  num_repeats = 1, crossing = NULL)
}
\arguments{
\item{datasets}{The datasets to be used in the evaluation.
Must be a named list consisting of dataset ids (character),
dynwrap::data_wrapper's, or functions that generate dynwrap::data_wrapper's.}

\item{methods}{The methods to be evaluated.
Must be a named list consisting of method_ids (character) or dynwrap::ti_wrapper's.}

\item{priors}{A list of lists. Each sublist contains a list of priors that each method is allowed to optionally use.
Check \code{\link[dynwrap:priors]{dynwrap::priors}} for a list of possible priors. Default priors given is "none"}

\item{num_repeats}{The number of times to repeat the evaluation.}

\item{crossing}{A data frame containing the combinations of methods, datasets and priors to be evaluated.
Must be a data frame containing the columns dataset_id, method_id, prior_id, repeat_ix and param_id}
}
\description{
Generate a benchmarking design
}
\examples{
\dontrun{
library(tibble)

datasets <- list(
  "synthetic/dyntoy/bifurcating_1",
  dyntoy::generate_dataset(id = "test1"),
  test2 = function() dyntoy::generate_dataset()
)

methods <- list(
  "comp1",
  dynmethods::ti_scorpius(),
  "tscan"
)

priors <- list(
  list(id = "none", method_id = c("comp1", "scorpius", "tscan")),
  list(id = "some", method_id = c("tscan"), set = c("start_id", "end_n"))
)

paramoptim_generate_design(
  datasets = datasets,
  methods = methods,
  priors = priors,
  num_repeats = 2
)
}
}
