% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_paramoptim.R
\name{paramoptim_qsub_fun}
\alias{paramoptim_qsub_fun}
\title{Helper function for paramoptim suite}
\usage{
paramoptim_qsub_fun(grid_i)
}
\arguments{
\item{grid_i}{paramoptim config index}
}
\description{
Helper function for paramoptim suite
}
