% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_paramoptim.R
\name{paramoptim_submit}
\alias{paramoptim_submit}
\title{A parameter optimisation suite}
\usage{
paramoptim_submit(task_ids, methods, timeout_paramoptim, max_memory_per_core,
  num_cores, metrics, num_repeats, num_iterations, num_init_params,
  local_output_folder, remote_output_folder, execute_before = NULL,
  verbose = FALSE)
}
\arguments{
\item{task_ids}{The ids of the tasks to be used in the evaluation.}

\item{methods}{A tibble of TI methods to use, see \code{\link[dynmethods]{get_ti_methods}}.}

\item{timeout_paramoptim}{The parameter optimisation timeout}

\item{max_memory_per_core}{The maximum amount of memory each core is allowed to use}

\item{num_cores}{The number of cores to use.}

\item{metrics}{Which metrics to evaluate; see \code{\link{calculate_metrics}} for a list of which metrics are available.}

\item{num_repeats}{The number of times to repeat the evaluation.}

\item{num_iterations}{The number of iterations to run.}

\item{num_init_params}{The number of initial parameters to evaluate.}

\item{local_output_folder}{A folder in which to output intermediate and final results.}

\item{remote_output_folder}{A folder in which to store intermediate results in a remote directory when using the PRISM package.}

\item{execute_before}{Shell commands to execute before running R.}

\item{verbose}{Whether or not to print extra information.}
}
\description{
A parameter optimisation suite
}
