% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formats.R
\name{pdf_manuscript}
\alias{pdf_manuscript}
\title{The manuscript pdf document}
\usage{

  pdf_manuscript(bibliography = paste0(dynbenchmark::get_dynbenchmark_folder(),
  "manuscript/assets/references.bib"),
  csl = paste0(dynbenchmark::get_dynbenchmark_folder(),
  "manuscript/assets/nature-biotechnology.csl"), render_changes = TRUE,
  ...)
}
\arguments{
\item{bibliography}{Bibliography location}

\item{csl}{Csl file location}

\item{render_changes}{Whether to export a *_changes.pdf as well}

\item{...}{Parameters for rmarkdown::pdf_document}
}
\description{
The manuscript pdf document
}
