% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_dataset.R
\name{preprocess_dataset}
\alias{preprocess_dataset}
\title{Preprocessing functionality for real datasets}
\usage{
preprocess_dataset(id = datasetpreproc_getid(), counts, milestone_network,
  grouping, root_milestone_id, cell_info = tibble(cell_id = rownames(counts)),
  feature_info = tibble(feature_id = colnames(counts)))
}
\arguments{
\item{id}{A unique identifier for the data}

\item{counts}{The counts with genes in columns and cells in rows}

\item{milestone_network}{A network of milestones.}

\item{grouping}{A grouping of the cells, can be a named vector or a dataframe with group_id (character) and cell_id (character)}

\item{root_milestone_id}{The id of the root milestone}

\item{cell_info}{Optional meta-information pertaining the cells.}

\item{feature_info}{Optional meta-information of the features, a data.frame with at least feature_id as column}
}
\description{
Preprocessing functionality for real datasets
}
