% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_helpers.R
\name{print_url}
\alias{print_url}
\title{Show a url}
\usage{
print_url(url, text = url, format = get_default_format())
}
\arguments{
\item{url}{The url}

\item{text}{The text}

\item{format}{The format}
}
\description{
Show a url
}
