% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prior_types}
\alias{prior_types}
\title{Metadata on prior types}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 2 columns.}
\usage{
prior_types
}
\description{
Metadata on prior types
}
\keyword{datasets}
