% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prior_usages}
\alias{prior_usages}
\title{Metadata on prior usage types}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 2 columns.}
\usage{
prior_usages
}
\description{
Metadata on prior usage types
}
\keyword{datasets}
