% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_priors.R
\docType{data}
\name{priors}
\alias{priors}
\title{The priors}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 1 columns.}
\usage{
priors
}
\description{
The priors
}
\keyword{datasets}
