% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_preprocessing.R
\name{process_raw_dataset}
\alias{process_raw_dataset}
\title{Process a raw real or synthetic dataset}
\usage{
process_raw_dataset(id, counts, milestone_network, grouping,
  root_milestone_id = NULL, cell_info = tibble(cell_id =
  rownames(counts)), feature_info = tibble(feature_id =
  colnames(counts)))
}
\arguments{
\item{id}{A unique identifier for the data. If \code{NULL}, a random string will be generated.}

\item{counts}{The counts with genes in columns and cells in rows}

\item{milestone_network}{A network of milestones.}

\item{grouping}{A grouping of the cells, can be a named vector or a dataframe with group_id (character) and cell_id (character)}

\item{root_milestone_id}{The root milestone, optional}

\item{cell_info}{Optional meta-information pertaining the cells.}

\item{feature_info}{Optional meta-information of the features, a data.frame with at least feature_id as column}
}
\description{
Process a raw real or synthetic dataset
}
