% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qc_categories}
\alias{qc_categories}
\title{Metadata on QC categories}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 3 columns.}
\usage{
qc_categories
}
\description{
Metadata on QC categories
}
\keyword{datasets}
