% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_ref.R
\name{ref}
\alias{ref}
\title{Refer to a figure or table}
\usage{
ref(ref_type, ref_id, suffix = "", anchor = FALSE,
  format = get_default_format())
}
\arguments{
\item{ref_type}{fig, sfig, table, ...}

\item{ref_id}{The identifier}

\item{suffix}{Adding something to the index}

\item{anchor}{Whether to anchor here}

\item{format}{The output format (html, latex, markdown, ...)}

\item{pattern}{What to use as pattern}
}
\description{
Refer to a figure or table
}
