% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_tribbles.R
\name{reformat_tribbles}
\alias{reformat_tribbles}
\title{Align tribbles in a file}
\usage{
reformat_tribbles(script_file, align = c("left", "center", "right"))
}
\arguments{
\item{script_file}{The file in which to align the tribble}

\item{align}{The alignment (left, center, right)}
}
\description{
The start and end of a tribble need to be marked with a commend \code{# tribble_start} and \code{# tribble_end}.
}
