% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_equations.R
\name{render_equations}
\alias{render_equations}
\title{Renders equations for github markdown}
\usage{
render_equations(x, format = get_default_format())
}
\arguments{
\item{x}{A character vector}

\item{format}{The format: html, markdown, latex, ...}
}
\description{
Renders equations for github markdown
}
