% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg_replacement.R
\name{replace}
\alias{replace}
\title{Replace
Replaces the rectangles in a given svg with the svgs contained in the replacer}
\usage{
replace(svg, replacer)
}
\arguments{
\item{svg}{The svg to replace}

\item{replacer}{Replacer dataframe}
}
\description{
Replace
Replaces the rectangles in a given svg with the svgs contained in the replacer
}
