% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_dataset_aggregation.R
\name{score_aggregation}
\alias{score_aggregation}
\title{Score milestone aggregation}
\usage{
score_aggregation(counts, cell_grouping)
}
\arguments{
\item{counts}{matrix of counts}

\item{cell_grouping}{data frame contain cell groups}
}
\description{
Score milestone aggregation
}
