% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_dataset_connectivity.R
\name{score_milestone_connectivity}
\alias{score_milestone_connectivity}
\title{Score milestone connectivity}
\usage{
score_milestone_connectivity(counts, cell_grouping, milestone_ids,
  milestone_network, perc = 0.1)
}
\arguments{
\item{counts}{counts matrix}

\item{cell_grouping}{data frame contain cell groups}

\item{milestone_ids}{vector of all milestones}

\item{milestone_network}{data frame contain the milestone network}

\item{perc}{Percentage of lowest distances to use between groups}
}
\description{
Score milestone connectivity
}
