% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_dataset_grouping.R
\name{score_milestone_grouping}
\alias{score_milestone_grouping}
\title{Score milestone grouping}
\usage{
score_milestone_grouping(counts, cell_grouping)
}
\arguments{
\item{counts}{counts matrix}

\item{cell_grouping}{data frame contain cell groups}
}
\description{
Score milestone grouping
}
