% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_dataset_milestone_transitions.R
\name{score_milestone_transitions}
\alias{score_milestone_transitions}
\title{Compare the known transitions of the milestone_network with all other possible transitions}
\usage{
score_milestone_transitions(counts_grouped, milestone_ids, milestone_network)
}
\arguments{
\item{counts_grouped}{grouped counts matrix}

\item{milestone_ids}{vector of all milestones}

\item{milestone_network}{data frame contain the milestone network}

\item{perc}{Percentage of lowest distances to use between groups}
}
\description{
Compare the known transitions of the milestone_network with all other possible transitions
}
