% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_ref.R
\name{setup_refs}
\alias{setup_refs}
\alias{setup_figs}
\alias{setup_tables}
\alias{setup_stables}
\title{Setup the text references}
\usage{
setup_refs()

setup_figs()

setup_tables()

setup_stables()
}
\description{
Setup the text references
}
