% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_ref.R
\name{setup_refs_globally}
\alias{setup_refs_globally}
\title{Setup the refs for a markdown document, globally!}
\usage{
setup_refs_globally()
}
\description{
Setup the refs for a markdown document, globally!
}
