% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trajectory_types.R
\name{simplify_trajectory_type}
\alias{simplify_trajectory_type}
\title{Convert trajectory type}
\usage{
simplify_trajectory_type(trajectory_types_oi)
}
\arguments{
\item{trajectory_types_oi}{trajectory types to simplify}
}
\description{
Convert trajectory type
}
