% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_utils.R
\name{strip_vglm}
\alias{strip_vglm}
\alias{strip_gam}
\title{Strip a vglm or gam model, only retaining those objects necessary for predicting}
\usage{
strip_vglm(model)

strip_gam(object, keep = "predict", ...)
}
\arguments{
\item{model}{The vglm model}
}
\description{
Strip a vglm or gam model, only retaining those objects necessary for predicting
}
\examples{
\dontrun{
d.AD <- data.frame(treatment = gl(3, 3),outcome = gl(3, 1, 9),counts = c(18,17,15,20,10,20,25,13,12))
model <- VGAM::vglm(counts ~ outcome + treatment, family = VGAM::poissonff, data = d.AD, trace = TRUE)

object.size(model)
object.size(strip_vglm(model))
1/as.numeric(object.size(strip_vglm(model))/object.size(model))
predict(model, d.AD)
predict(strip_vglm(model), d.AD)
}
model <- scaling_results$models$model_time[[1]]
model2 <- strip_vglm(model)
model2@family \%>\% maps4(pryr::object_size) \%>\% unlist() \%>\% sort()
model2@family@linkinv \%>\% pryr::object_size()
}
