% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_labels.R
\name{tag_first}
\alias{tag_first}
\title{tag the first plot of an assemble}
\usage{
tag_first(x, tag)
}
\arguments{
\item{x}{A ggassemble}

\item{tag}{The tag}
}
\description{
tag the first plot of an assemble
}
