% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_themes.R
\docType{data}
\name{theme_empty_x_axis}
\alias{theme_empty_x_axis}
\title{Themes which make certain parts of the plot empty}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 4.}
\usage{
theme_empty_x_axis
}
\description{
Themes which make certain parts of the plot empty
}
\keyword{datasets}
