% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataraw_method_characterisation.R
\docType{data}
\name{topinf}
\alias{topinf}
\title{Topology inference}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 3 columns.}
\usage{
topinf
}
\description{
Topology inference
}
\keyword{datasets}
