% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataraw_method_characterisation.R
\docType{data}
\name{topinf_colours}
\alias{topinf_colours}
\title{Topology inference colours}
\format{An object of class \code{character} of length 3.}
\usage{
topinf_colours
}
\description{
Topology inference colours
}
\keyword{datasets}
