% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{topinf_types}
\alias{topinf_types}
\title{Metadata on topology inference types}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 4 columns.}
\usage{
topinf_types
}
\description{
Metadata on topology inference types
}
\keyword{datasets}
