% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataraw_trajectory_types.R
\docType{data}
\name{trajectory_type_dag}
\alias{trajectory_type_dag}
\title{Trajectory type DAG}
\format{An object of class \code{tbl_graph} (inherits from \code{igraph}) of length 10.}
\usage{
trajectory_type_dag
}
\description{
Trajectory type DAG
}
\keyword{datasets}
