% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_trajectory_types.R
\docType{data}
\name{trajectory_types}
\alias{trajectory_types}
\title{Trajectory types}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17 rows and 5 columns.}
\usage{
trajectory_types
}
\description{
Trajectory types
}
\keyword{datasets}
