% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataraw_trajectory_types.R
\docType{data}
\name{trajectory_types_simplified}
\alias{trajectory_types_simplified}
\title{Simplified traj types}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 6 columns.}
\usage{
trajectory_types_simplified
}
\description{
Simplified traj types
}
\keyword{datasets}
