% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formats.R
\name{word_manuscript}
\alias{word_manuscript}
\title{A nested github markdown document}
\usage{
word_manuscript(bibliography = raw_file("references.bib",
  "12-manuscript"), csl = raw_file("nature-biotechnology.csl",
  "12-manuscript"), ...)
}
\arguments{
\item{bibliography}{Bibliography location}

\item{csl}{Csl file location}

\item{...}{Parameters for rmarkdown::github_document}
}
\description{
A nested github markdown document
}
