% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_ref.R
\name{add_table}
\alias{add_table}
\alias{add_stable}
\title{Add a table}
\usage{
add_table(table, ref_id, caption_main, caption_text = "",
  format = get_default_format())

add_stable(table, ref_id, caption_main, caption_text = "",
  format = get_default_format())
}
\arguments{
\item{table}{Either a tibble, a path to a table or a named list with latex, html and markdown kables}

\item{ref_id}{The identifier}

\item{caption_main}{Caption title}

\item{caption_text}{Caption text}

\item{format}{The format, in html or latex}
}
\description{
Add a table
}
