% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_benchmark_fetch_results.R
\name{benchmark_bind_results}
\alias{benchmark_bind_results}
\title{Gather and bind the results of the benchmark jobs}
\usage{
benchmark_bind_results(load_models = FALSE, filter_fun = NULL,
  local_output_folder = derived_file("suite"))
}
\arguments{
\item{load_models}{Whether or not to load the models as well.}

\item{filter_fun}{A function with which to filter the data as it is being read from files. Function must take a single data frame as input and return a filtered data frame as a result.}

\item{local_output_folder}{A folder in which to output intermediate and final results.}
}
\description{
Gather and bind the results of the benchmark jobs
}
