% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_benchmark_fetch_results.R
\name{benchmark_bind_results}
\alias{benchmark_bind_results}
\title{Gather and bind the results of the benchmark jobs}
\usage{
benchmark_bind_results(load_models = FALSE, experiment_id = NULL,
  filter_fun = NULL)
}
\arguments{
\item{load_models}{Whether or not to load the models as well.}

\item{experiment_id}{The experiment_id, defaults to the current one}

\item{filter_fun}{A function with which to filter the data as it is being read from files. Function takes a single data frame as input and returns a filtered data frame as a result.}
}
\description{
Gather and bind the results of the benchmark jobs
}
