% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_benchmark_submit.R
\name{benchmark_qsub_fun}
\alias{benchmark_qsub_fun}
\title{Helper function for benchmark suite}
\usage{
benchmark_qsub_fun(i)
}
\arguments{
\item{i}{Row of a design dataframe}
}
\description{
Helper function for benchmark suite
}
