% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_benchmark.R
\name{benchmark_qsub_fun}
\alias{benchmark_qsub_fun}
\title{Helper function for benchmark suite}
\usage{
benchmark_qsub_fun(design_row)
}
\arguments{
\item{grid_i}{Benchmark config index}
}
\description{
Helper function for benchmark suite
}
