% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formats.R
\name{common_dynbenchmark_format}
\alias{common_dynbenchmark_format}
\title{Common dynbenchmark format}
\usage{
common_dynbenchmark_format(format,
  bibliography = raw_file("references.bib", "12-manuscript"),
  csl = raw_file("nature-biotechnology.csl", "12-manuscript"),
  reference_section_title = "References")
}
\arguments{
\item{format}{The format on which to apply common changes}

\item{bibliography}{Bibliography location}

\item{csl}{Csl file location}

\item{reference_section_title}{The title of the reference section}
}
\description{
Common dynbenchmark format
}
