% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_datasets.R
\name{dataset_preprocessing}
\alias{dataset_preprocessing}
\alias{get_dataset_preprocessing_id}
\alias{dataset_source_file}
\alias{dataset_raw_file}
\alias{save_raw_dataset}
\alias{dataset_file}
\alias{save_dataset}
\title{Helper functions for creating new datasets}
\usage{
dataset_preprocessing(dataset_id)

get_dataset_preprocessing_id()

dataset_source_file(filename = "", dataset_id = NULL, remote = FALSE)

dataset_raw_file(dataset_id)

save_raw_dataset(dataset, dataset_id = get_dataset_preprocessing_id())

dataset_file(filename = "", dataset_id = NULL, remote = FALSE)

save_dataset(dataset, dataset_id = NULL, lazy_load = TRUE)
}
\arguments{
\item{dataset_id}{The ID of the dataset to be used}

\item{filename}{Custom filename}

\item{remote}{The remote to access, "" if working locally, TRUE if using default qsub_config remote}

\item{dataset}{Dataset object to save}

\item{lazy_load}{Whether or not to allow for lazy loading of large objects in the dataset}
}
\description{
Helper functions for creating new datasets
}
