% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dataset_preprocessing.R
\name{dataset_preprocessing}
\alias{dataset_preprocessing}
\alias{datasetpreproc_getid}
\alias{dataset_preproc_file}
\alias{dataset_file}
\alias{save_dataset}
\title{Helper functions for creating new datasets}
\usage{
dataset_preprocessing(dataset_id)

datasetpreproc_getid()

dataset_preproc_file(filename = "", dataset_id = NULL, glue = NULL,
  relative = FALSE)

dataset_file(filename = "", dataset_id = NULL, glue = NULL,
  relative = FALSE)

save_dataset(dataset, dataset_id = NULL)
}
\arguments{
\item{dataset_id}{The ID of the dataset to be used}

\item{filename}{Custom filename}

\item{glue}{if specified, glue::glue this string and override filename.}

\item{relative}{Whether or not to output relative paths}

\item{dataset}{Dataset object to save}
}
\description{
Helper functions for creating new datasets
}
