% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_loading.R
\name{dataset_preprocessing}
\alias{dataset_preprocessing}
\alias{dataset_preproc_file}
\alias{dataset_file}
\alias{save_dataset}
\title{Helper function for creating new datasets}
\usage{
dataset_preprocessing(prefix, dataset_id)

dataset_preproc_file(...)

dataset_file(...)

save_dataset(dataset)
}
\arguments{
\item{prefix}{Dataset prefix}

\item{dataset_id}{Dataset id}

\item{...}{filenames}
}
\description{
Helper function for creating new datasets
}
