% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_dataset_preprocessing.R
\name{dataset_preprocessing}
\alias{dataset_preprocessing}
\alias{datasetpreproc_getprefix}
\alias{datasetpreproc_getid}
\alias{dataset_preproc_file}
\alias{dataset_file}
\alias{save_dataset}
\title{Helper functions for creating new datasets}
\usage{
dataset_preprocessing(prefix, dataset_id)

datasetpreproc_getprefix()

datasetpreproc_getid()

dataset_preproc_file(..., prefix = NULL, dataset_id = NULL)

dataset_file(..., prefix = NULL, dataset_id = NULL)

save_dataset(dataset, prefix = NULL, dataset_id = NULL)
}
\arguments{
\item{prefix}{Dataset prefix}

\item{dataset_id}{Dataset id}

\item{...}{filenames}

\item{dataset}{Dataset object to save}
}
\description{
Helper functions for creating new datasets
}
