% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_datasets.R
\name{download_dataset_source_file}
\alias{download_dataset_source_file}
\title{Download a file and return its location path}
\usage{
download_dataset_source_file(filename, url, id = NULL)
}
\arguments{
\item{filename}{What name to give to the file}

\item{url}{The url of the file to download}

\item{id}{An optional id}
}
\description{
Download a file and return its location path
}
