% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_package.R
\docType{package}
\name{dynbenchmark}
\alias{dynbenchmark}
\alias{dynbenchmark-package}
\title{Easy installing and loading of the dynverse}
\description{
The 'dynverse' is a set of packages with which to
evaluate trajectory inference methods for dynamic processes.
}
\details{
The dynbenchmark package contains the code to generate/preprocess datasets, evaluate TI methods on these datasets, and finally wrap all results together in a scientific paper.
The code within this package can be run locally, but can also be run within a docker environment, for which the scripts are contained in the `scripts` folder.
}
