% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_setup.R
\name{experiment}
\alias{experiment}
\alias{derived_file}
\alias{raw_file}
\alias{figure_file}
\alias{result_file}
\title{Helper function for controlling experiments}
\usage{
experiment(experiment_id)

derived_file(...)

raw_file(...)

figure_file(...)

result_file(...)
}
\arguments{
\item{experiment_id}{id for the experiment}

\item{...}{filenames}
}
\description{
Helper function for controlling experiments
}
\examples{
\dontrun{
experiment("test_plots")

data <- matrix(runif(200), ncol = 2)
pdf(figure_file("testplot.pdf"), 5, 5)
plot(data)
dev.off()
}
}
