% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formats.R
\name{fix_relative_paths}
\alias{fix_relative_paths}
\title{process relative paths to links & figures
match every link and check whether it is an absolute path (= starts with / or http) or not
every relative path gets updated}
\usage{
fix_relative_paths(knit, folder)
}
\arguments{
\item{knit}{Character vector}

\item{folder}{The relative folder}
}
\description{
process relative paths to links & figures
match every link and check whether it is an absolute path (= starts with / or http) or not
every relative path gets updated
}
\examples{
knit <- c(
"hshlkjdsljkfdhg [i am a absolute path](/pompompom/dhkjhlkj/) kjfhlqkjsdhlkfjqsdf",
"hshlkjdsljkfdhg [i am a relative path](pompompom/dhkjhlkj/) kjfhlqkjsdhlkfjqsdf",
"<img src = \\"heyho/heyho\\">",
"<img src = \\"/heyho/heyho\\">"
)
dynbenchmark:::fix_relative_paths(knit, "IT WORKED :)")
}
