% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formats.R
\name{fix_relative_paths}
\alias{fix_relative_paths}
\title{Process relative paths to links & figures
First extract every link, determine whether it is a relative path and if yes, add folder to the front}
\usage{
fix_relative_paths(knit, folder)
}
\arguments{
\item{knit}{Character vector}

\item{folder}{The relative folder}
}
\description{
Process relative paths to links & figures
First extract every link, determine whether it is a relative path and if yes, add folder to the front
}
\examples{
knit <- c(
"hshlkjdsljkfdhg [i am a absolute path](/pompompom/dhkjhlkj/) kjfhlqkjsdhlkfjqsdf",
"hshlkjdsljkfdhg [i am a relative path](pompompom/dhkjhlkj/) kjfhlqkjsdhlkfjqsdf",
"<img src = \\"heyho/heyho\\">",
"<img src = \\"/heyho/heyho\\">"
)
dynbenchmark:::fix_relative_paths(knit, "IT WORKED :)")
}
