% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funky_heatmap.R
\name{funky_heatmap}
\alias{funky_heatmap}
\title{Overview heatmap plotting}
\usage{
funky_heatmap(data, column_info, column_groups, row_info, row_groups,
  palettes, scale_column = TRUE, add_timestamp = TRUE)
}
\arguments{
\item{data}{data}

\item{add_timestamp}{Whether or not to add a timestamp at the bottom}

\item{metric_ord}{metric_ord}

\item{method_ord}{method_ord}
}
\description{
TODO: params need to be described
}
