% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funky_heatmap.R
\name{funky_heatmap}
\alias{funky_heatmap}
\title{Overview heatmap plotting}
\usage{
funky_heatmap(data, column_info, column_groups, row_info, row_groups,
  palettes, scale_column = TRUE, add_timestamp = TRUE)
}
\arguments{
\item{data}{data}

\item{column_info}{a}

\item{column_groups}{b}

\item{row_info}{c}

\item{row_groups}{d}

\item{palettes}{e}

\item{scale_column}{f}

\item{add_timestamp}{Whether or not to add a timestamp at the bottom}
}
\description{
TODO: params need to be described
}
