% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_benchmark.R
\name{generate_benchmark_design}
\alias{generate_benchmark_design}
\title{Generate a benchmarking design}
\usage{
generate_benchmark_design(datasets, methods, parameters = NULL,
  give_priors = NULL, num_repeats = 1)
}
\arguments{
\item{datasets}{The datasets to be used in the evaluation.
Can be a character vector (the dataset ids), a list of
dynwrap datasets, or a list of functions each of which will generate
a dynwrap dataset.}

\item{methods}{The methods to be evaluated.
Can be a character vector (the method ids), or a list of
dynwrap TI methods.}

\item{parameters}{A named list containing data frames of the parameters to evaluate.
The names of the list must be present in method_ids.
The data frames must be of format \code{data_frame(paramset_id = "set1", param1 = "a", param2 = 2.0)}.}

\item{give_priors}{All the priors a method is allowed to receive. Must be a subset of: \code{"start_milestones"},
\code{"start_id"}, \code{"end_milestones"}, \code{"end_id"}, \code{"groups_id"} and \code{"groups_network"}}

\item{num_repeats}{The number of times to repeat the evaluation.}
}
\description{
Generate a benchmarking design
}
\examples{
library(tibble)
generate_benchmark_design(
  datasets = c("toy/bifurcating_1", "toy/bifurcating_2"),
  method_ids = c("angle", "scorpius", "tscan"),
  parameters = list(
    scorpius = tibble(paramset_ix = 1),
    tscan = tibble(paramset_ix = 1:3, clusternum_lower = 4:6, clusternum_upper = 18:20)
 ),
  give_priors = NULL,
  num_repeats = 2
)
}
