% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formats.R
\name{github_markdown_nested}
\alias{github_markdown_nested}
\title{A nested github markdown document}
\usage{
github_markdown_nested(bibliography = raw_file("references.bib",
  "12-manuscript"), csl = raw_file("nature-biotechnology.csl",
  "12-manuscript"), ...)
}
\arguments{
\item{bibliography}{Bibliography location}

\item{csl}{Csl file location}

\item{...}{Parameters for rmarkdown::github_document}
}
\description{
A nested github markdown document
}
