% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formats.R
\name{knit_nest}
\alias{knit_nest}
\title{Knit a child, and add an extra level of headings + fix relative paths. Can be used both for latex and markdown output formats}
\usage{
knit_nest(file, levels = 1)
}
\arguments{
\item{file}{File to knit, can also be a directory in which case the README.Rmd will be knit}

\item{levels}{The number of levels to add onto section headings}
}
\description{
Knit a child, and add an extra level of headings + fix relative paths. Can be used both for latex and markdown output formats
}
