% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_labels.R
\name{label_metric}
\alias{label_metric}
\alias{label_metrics}
\title{Label the metrics}
\usage{
label_metric(metric_id, format = get_default_metric_format(),
  parse = FALSE)

label_metrics(metric_ids, ...)
}
\arguments{
\item{metric_id}{metric id}

\item{format}{The format of the label to return, can be plotmath, latex, long_name or metric_id}

\item{parse}{Whether to parse the label into an expression}

\item{metric_ids}{The metric ids}

\item{...}{Extra parameters for label_metric}
}
\description{
Label the metrics
}
