% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_labels.R
\name{label_metric}
\alias{label_metric}
\alias{label_metrics}
\title{Label the metrics}
\usage{
label_metric(metric_id, label_type = c("plotmath", "latex", "long_name"),
  parse = FALSE)

label_metrics(metric_ids, ...)
}
\arguments{
\item{metric_id}{metric id}

\item{parse}{Whether to parse the label into an expression}

\item{metric_ids}{The metric ids}

\item{...}{Extra parameters for label_metric}
}
\description{
Label the metrics
}
