% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_labels.R
\name{label_simple_trajectory_types}
\alias{label_simple_trajectory_types}
\title{Label trajectory types simplified}
\usage{
label_simple_trajectory_types(trajectory_types_oi)
}
\arguments{
\item{trajectory_types_oi}{Trajectory types}
}
\description{
Label trajectory types simplified
}
