% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_labels.R
\docType{data}
\name{labels}
\alias{labels}
\title{Common manual labeling}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30 rows and 3 columns.}
\usage{
labels
}
\description{
Common manual labeling
}
\keyword{datasets}
