% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labels}
\alias{labels}
\title{Common manual labelling}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36 rows and 3 columns.}
\usage{
labels
}
\description{
Common manual labelling
}
\keyword{datasets}
