% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_loading.R
\name{load_dataset}
\alias{load_dataset}
\title{Loading a dataset after it has been preprocessed}
\usage{
load_dataset(prefix, dataset_id)
}
\arguments{
\item{prefix}{Dataset prefix}

\item{dataset_id}{Dataset id}
}
\description{
Loading a dataset after it has been preprocessed
}
