% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_datasets.R
\name{list_datasets}
\alias{list_datasets}
\alias{load_dataset}
\alias{load_datasets}
\title{Load datasets}
\usage{
list_datasets()

load_dataset(dataset_id, as_tibble = FALSE)

load_datasets(dataset_ids = list_datasets()$dataset_id, as_tibble = TRUE)
}
\arguments{
\item{dataset_id}{The ID of the dataset to be used}

\item{as_tibble}{Return the datasets as a tibble or as a list of datasets?}

\item{dataset_ids}{Character vector of dataset identifiers}
}
\description{
Load datasets
}
