% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_methods.R
\name{load_methods}
\alias{load_methods}
\title{Load the information on the methods}
\usage{
load_methods(default_filter = method_source == "tool" | method_id \%in\%
  c("angle", "comp1", "mst"), ...)
}
\arguments{
\item{default_filter}{Default filtering for the benchmark.}

\item{...}{Other filters for the method tibble.}
}
\description{
Load the information on the methods
}
