% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_paramoptim.R
\name{make_obj_fun}
\alias{make_obj_fun}
\title{Used for wrapping an evaluation function around a TI method}
\usage{
make_obj_fun(method, metrics, extra_metrics, noisy = FALSE, verbose = FALSE)
}
\arguments{
\item{method}{The method to evaluate.}

\item{metrics}{which metrics to evaluate:
\enumerate{
  \item Spearman correlation of geodesic distances: \code{"correlation"}
  \item Edge flip score: \code{"edge_flip"}
  \item RF MSE: \code{"rf_mse"}, \code{"rf_rsq"}
  \item Similarity in feature importance: \code{"featureimp_cor"}
  \item Custom metric function. Format: \code{function(dataset, model) { 1 }}
}}

\item{extra_metrics}{Extra metrics to calculate but not evaluate with.}

\item{noisy}{Whether or not the metric is noisy or not}

\item{verbose}{Whether or not to print extra information output}
}
\description{
Used for wrapping an evaluation function around a TI method
}
