% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_paramoptim.R
\name{make_obj_fun}
\alias{make_obj_fun}
\title{Used for wrapping an evaluation function around a TI method}
\usage{
make_obj_fun(method, metrics, noisy = FALSE, verbose = FALSE)
}
\arguments{
\item{method}{One or more methods. Must be one of:
\itemize{
\item{an object or list of ti_... objects (eg. \code{dynmethods::ti_comp1()}),}
\item{a character vector containing the names of methods to execute (e.g. \code{"scorpius"}),}
\item{a character vector containing dockerhub repositories (e.g. \code{dynverse/paga}), or}
\item{a dynguidelines data frame.}
}}

\item{metrics}{which metrics to evaluate:
\enumerate{
  \item Spearman correlation of geodesic distances: \code{"correlation"}
  \item Edge flip score: \code{"edge_flip"}
  \item Isomorphic: \code{"isomorphic"}
  \item RF MSE: \code{"rf_mse"}, \code{"rf_rsq"}
  \item Similarity in feature importance: \code{"featureimp_cor"}, \code{"featureimp_wcor"}, \code{"featureimp_ks"}, \code{"featureimp_wilcox"}
  \item Overlap between branches: \code{"F1_branches"}
  \item Custom metric function. Format: \code{function(dataset, model) { 1 }}
}}

\item{noisy}{Whether or not the metric is noisy or not}

\item{verbose}{Whether or not to print extra information output}
}
\description{
Used for wrapping an evaluation function around a TI method
}
