% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_paramoptim_submit.R
\name{make_obj_fun}
\alias{make_obj_fun}
\title{Used for wrapping an evaluation function around a TI method}
\usage{
make_obj_fun(method, metrics, give_priors, noisy = FALSE,
  verbose = FALSE)
}
\arguments{
\item{method}{One or more methods. Must be one of:
\itemize{
\item{an object or list of ti_... objects (eg. \code{dynmethods::ti_comp1()}),}
\item{a character vector containing the names of methods to execute (e.g. \code{"scorpius"}),}
\item{a character vector containing dockerhub repositories (e.g. \code{dynverse/paga}), or}
\item{a dynguidelines data frame.}
}}

\item{metrics}{Which metrics to evaluate. Check \code{dyneval::metrics} for a list of possible metrics.
Passing a custom metric function with format \code{function(dataset, model) { 1 }} is also supported. The name of this function within the list will be used as the name of the metric.}

\item{give_priors}{All the priors a method is allowed to receive.
Must be a subset of all available priors (\code{\link[dynwrap:priors]{priors}}).}

\item{noisy}{Whether or not the metric is noisy or not}

\item{verbose}{Whether or not to print extra information output}
}
\description{
Used for wrapping an evaluation function around a TI method
}
