% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_metrics.R
\docType{data}
\name{metrics}
\alias{metrics}
\title{Common metric information}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.}
\usage{
metrics
}
\description{
Common metric information
}
\keyword{datasets}
