% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_non_inclusion_reasons.R
\docType{data}
\name{non_inclusion_reasons}
\alias{non_inclusion_reasons}
\title{Non inclusion reasons}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 3 columns.}
\usage{
non_inclusion_reasons
}
\description{
Non inclusion reasons
}
\keyword{datasets}
