% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_paramoptim_fetch_results.R
\name{paramoptim_bind_results}
\alias{paramoptim_bind_results}
\title{Gather and bind the results of the benchmark jobs}
\usage{
paramoptim_bind_results(filter_fun = NULL,
  local_output_folder = derived_file("suite"))
}
\arguments{
\item{filter_fun}{A function with which to filter the data as it is being read from files. Function must take a single data frame as input and return a filtered data frame as a result.}

\item{local_output_folder}{A folder in which to output intermediate and final results.}
}
\description{
Gather and bind the results of the benchmark jobs
}
