% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_paramoptim_fetch_results.R
\name{paramoptim_fetch_results}
\alias{paramoptim_fetch_results}
\title{Fetch the results of the benchmark jobs from the cluster.}
\usage{
paramoptim_fetch_results(remote = NULL,
  local_output_folder = derived_file("suite"), impatient = FALSE)
}
\arguments{
\item{remote}{The host at which to check for running jobs.
If \code{NULL}, each qsub handle will be checked individually.
If \code{TRUE}, the default qsub handle will be used.}

\item{local_output_folder}{A folder in which to output intermediate and final results.}

\item{impatient}{If true, then the results will be fetched regardless of whether the optimisation has finished.}
}
\description{
Fetch the results of the benchmark jobs from the cluster.
}
