% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_paramoptim_submit.R
\name{paramoptim_qsub_fun}
\alias{paramoptim_qsub_fun}
\title{Helper function for parameter optimisation suite}
\usage{
paramoptim_qsub_fun(i)
}
\arguments{
\item{i}{Row of a design dataframe}
}
\description{
Helper function for parameter optimisation suite
}
