% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_paramoptim_submit.R
\name{paramoptim_submit}
\alias{paramoptim_submit}
\title{A param optim suite with which to run all the methods on the different datasets}
\usage{
paramoptim_submit(design, metrics = "correlation",
  qsub_grouping = "{method_id}/{prior_id}/{repeat_ix}",
  qsub_params = list(timeout = 3600, memory = "10G", num_cores = 8,
  num_iterations = 100, num_init_params = 39), verbose = TRUE,
  seed = function() random_seed(),
  local_output_folder = derived_file("suite"),
  remote_output_folder = derived_file("suite", remote = TRUE))
}
\arguments{
\item{design}{Design tibble of the experiment, created by \code{\link[=paramoptim_generate_design]{paramoptim_generate_design()}}.}

\item{metrics}{Which metrics to evaluate; see \code{\link[=calculate_metrics]{calculate_metrics()}} for a list of which metrics are available.}

\item{qsub_grouping}{A character used to partition the design into separate jobs. Any of the column names
in \code{design$crossing} is allowed to be used. This string will later be parsed by \code{\link[glue:glue]{glue::glue()}}.}

\item{qsub_params}{A list used to define execution parameters for each row in the design tibble.
\itemize{
\item \code{memory} defines the amount of memory,
\item \code{timeout} defines the maximum wall time,
\item \code{num_cores} defines the number of cores
\item \code{num_iterations} defines the number of mlrmbo iterations
\item \code{num_init_params} defines the number of parameters in the iteration 0.
}

Optionally, a function in the format \code{function(XXX, YYY, ...) { ZZZ }} is possible, where XXX and YYY
are equal to the groups defined by \code{qsub_grouping} (default \code{method_id} and \code{param_id}),
and ZZZ is equal to some logic which always produces a \code{list(memory = ..., timeout = ..., num_cores ...)}.}

\item{verbose}{Whether or not to print extra information.}

\item{seed}{A seed for the parameter optimisation}

\item{local_output_folder}{A folder in which to output intermediate and final results.}

\item{remote_output_folder}{A folder in which to store intermediate results in a remote directory when using the qsub package.}
}
\description{
A param optim suite with which to run all the methods on the different datasets
}
\examples{
\dontrun{
library(tibble)

datasets <- c("synthetic/dyntoy/bifurcating_1", "synthetic/dyntoy/bifurcating_2")
methods <- dynmethods::methods$method_id[1:3]

design <- paramoptim_generate_design(
  datasets = datasets,
  methods = methods
)

paramoptim_submit(
  design = design,
  metrics = c("correlation", "rf_mse"),
  qsub_grouping = "{method_id}/{replicate_ix}",
  qsub_params = function(method_id, replicate_ix) {
    params <- lst(memory = "10G", timeout = 3600)
    if (method_id == "scorpius") params$memory <- "5G"
    params
  }
)
}
}
