% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formats.R
\name{pdf_manuscript}
\alias{pdf_manuscript}
\title{The manuscript pdf document}
\usage{
pdf_manuscript(bibliography = raw_file("references.bib",
  "12-manuscript"), csl = raw_file("nature-biotechnology.csl",
  "12-manuscript"), render_changes = TRUE, toc = FALSE,
  linenumbers = FALSE, sty_header = NULL, sty_beforebody = NULL,
  sty_afterbody = NULL,
  reference_section_title = reference_section_title, ...)
}
\arguments{
\item{bibliography}{Bibliography location}

\item{csl}{Csl file location}

\item{render_changes}{Whether to export a *_changes.pdf as well}

\item{toc}{Whether to include a table of contents}

\item{linenumbers}{Whether to add linenumbers to the left side}

\item{sty_header}{Extra sty files to be included in the header of the document}

\item{sty_beforebody}{Extra sty files to be included before the body of the document}

\item{sty_afterbody}{Extra sty files to be included after the body of the document}

\item{reference_section_title}{The title of the reference section}

\item{...}{Parameters for rmarkdown::pdf_document}
}
\description{
The manuscript pdf document
}
