% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formats.R
\name{pdf_supplementary_note}
\alias{pdf_supplementary_note}
\title{A supplementary note pdf document}
\usage{
pdf_supplementary_note(...)
}
\arguments{
\item{...}{Parameters for rmarkdown::pdf_document}
}
\description{
A supplementary note pdf document
}
