% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_trajectory_types.R
\name{plot_trajectory_types}
\alias{plot_trajectory_types}
\title{Plot a trajectory type}
\usage{
plot_trajectory_types(plot = ggplot() + theme_void(), trajectory_types,
  xmins = 0, xmaxs = 1, ymins = 0, ymaxs = 1,
  node_colours = NULL, edge_colours = NULL, size = 4,
  geom = c("point", "circle"), circ_size = ifelse(geom == "circle",
  0.05, 0))
}
\arguments{
\item{plot}{Existing plot}

\item{trajectory_types}{One or more trajectory types}

\item{xmins}{Minimal x bound (length 1 or length of \code{trajectory_types})}

\item{xmaxs}{Maximal y bound (likewise)}

\item{ymins}{Minimal y bound (likewise)}

\item{ymaxs}{Maximal y bound (likewise)}

\item{node_colours}{The colours of the nodes.
If \code{NULL}, the colours provided by \code{dynwrap::trajectory_types} will be used (recommended).
Otherwise, \code{node_colours} is expected to be of length 1 or length of \code{trajectory_types}.}

\item{edge_colours}{The colours of the edges.
If \code{NULL}, the edges will be coloured black.
Otherwise, \code{edge_colours} is expected to be of length 1 or length of \code{trajectory_types}.}

\item{size}{Size of the arrows and nodes}

\item{geom}{Which underlying geom to use. When using \code{"point"}, it is expected to also be
using \code{scale_colour_identity()} with your plot. The nodes will go slightly outside the (x, y) boundaries,
because the size of the points depend on the plot size. When using \code{"circle"}, it is expected to be using
\code{scale_fill_identity()} and \code{coord_equal()}, otherwise the nodes might be elliptical.`}

\item{circ_size}{Size of nodes, only used when \code{geom == "circle"}.}
}
\description{
Plot a trajectory type
}
