% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_trajectory_types.R
\name{plot_trajectory_types}
\alias{plot_trajectory_types}
\title{Plot a trajectory type}
\usage{
plot_trajectory_types(plot = ggplot() + theme_void(), trajectory_types,
  xmins = 0, xmaxs = 1, ymins = 0, ymaxs = 1, size = 4)
}
\arguments{
\item{plot}{Existing plot}

\item{trajectory_types}{One or more trajectory types}

\item{xmins}{Minimal x bound}

\item{xmaxs}{Maximal y bound}

\item{ymins}{Minimal y bound}

\item{ymaxs}{Maximal y bound}

\item{size}{Size of the arrows and dots}
}
\description{
Plot a trajectory type
}
