% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qc_applications}
\alias{qc_applications}
\title{Metadata on QC applications}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 1 columns.}
\usage{
qc_applications
}
\description{
Metadata on QC applications
}
\keyword{datasets}
