% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trajectory_type_dag}
\alias{trajectory_type_dag}
\title{A DAG of trajectory types}
\format{An object of class \code{tbl_graph} (inherits from \code{igraph}) of length 10.}
\usage{
trajectory_type_dag
}
\description{
A DAG of trajectory types
}
\keyword{datasets}
