% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trajectory_types}
\alias{trajectory_types}
\title{Metadata on the trajectory types}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19 rows and 7 columns.}
\usage{
trajectory_types
}
\description{
Metadata on the trajectory types
}
\keyword{datasets}
