% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trajectory_types_simplified}
\alias{trajectory_types_simplified}
\title{Metadata on simplified trajectory types}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11 rows and 3 columns.}
\usage{
trajectory_types_simplified
}
\description{
Metadata on simplified trajectory types
}
\keyword{datasets}
