% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formats.R
\name{word_manuscript}
\alias{word_manuscript}
\title{A nested github markdown document}
\usage{

  word_manuscript(bibliography = fs::path_abs(paste0(dynbenchmark::get_dynbenchmark_folder(),
  "manuscript/assets/references.bib")),
  csl = fs::path_abs(paste0(dynbenchmark::get_dynbenchmark_folder(),
  "manuscript/assets/nature-biotechnology.csl")), ...)
}
\arguments{
\item{bibliography}{Bibliography location}

\item{csl}{Csl file location}

\item{...}{Parameters for rmarkdown::github_document}
}
\description{
A nested github markdown document
}
